DPR2-1148 split configured API service
