+++ src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -11,7 +11,8 @@ uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.ProductDefinitionRep
 uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.RedShiftSummaryTableHelper
 uk.gov.justice.digital.hmpps.digitalprisonreportinglib.security.CaseloadProviderAutoConfig
 uk.gov.justice.digital.hmpps.digitalprisonreportinglib.security.ResourceServerConfiguration
-uk.gov.justice.digital.hmpps.digitalprisonreportinglib.service.ConfiguredApiService
+uk.gov.justice.digital.hmpps.digitalprisonreportinglib.service.SyncDataApiService
+uk.gov.justice.digital.hmpps.digitalprisonreportinglib.service.AsyncDataApiService
 uk.gov.justice.digital.hmpps.digitalprisonreportinglib.service.ReportDefinitionMapper
 uk.gov.justice.digital.hmpps.digitalprisonreportinglib.service.ReportDefinitionSummaryMapper
 uk.gov.justice.digital.hmpps.digitalprisonreportinglib.service.ReportDefinitionService
