+++ src/test/kotlin/uk/gov/justice/digital/hmpps/digitalprisonreportinglib/service/SyncDataApiServiceTest.kt
@@ -7,38 +7,31 @@ import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
-import org.junit.jupiter.params.provider.CsvSource
 import org.junit.jupiter.params.provider.ValueSource
 import org.mockito.kotlin.any
-import org.mockito.kotlin.anyOrNull
 import org.mockito.kotlin.mock
 import org.mockito.kotlin.times
 import org.mockito.kotlin.verify
 import org.mockito.kotlin.verifyNoInteractions
 import org.mockito.kotlin.whenever
-import org.springframework.jdbc.UncategorizedSQLException
 import org.springframework.security.core.authority.SimpleGrantedAuthority
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.config.DefinitionGsonConfig
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.controller.DataApiSyncController.FiltersPrefix.RANGE_FILTER_END_SUFFIX
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.controller.DataApiSyncController.FiltersPrefix.RANGE_FILTER_START_SUFFIX
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.controller.model.Count
-import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.AthenaApiRepository
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.ConfiguredApiRepository
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.ConfiguredApiRepository.Filter
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.DatasetHelper
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.IsoLocalDateTimeTypeAdaptor
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.JsonFileProductDefinitionRepository
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.ProductDefinitionRepository
-import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.RedshiftDataApiRepository
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.RepositoryHelper.Companion.EXTERNAL_MOVEMENTS_PRODUCT_ID
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.RepositoryHelper.FilterType.BOOLEAN
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.RepositoryHelper.FilterType.DATE_RANGE_END
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.RepositoryHelper.FilterType.DATE_RANGE_START
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.RepositoryHelper.FilterType.DYNAMIC
-import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.RepositoryHelper.FilterType.STANDARD
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.model.Dataset
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.model.Datasource
-import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.model.FilterType
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.model.MetaData
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.model.ParameterType
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.model.ProductDefinition
@@ -56,23 +49,15 @@ import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.model.policye
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.model.policyengine.Policy.PolicyResult.POLICY_PERMIT
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.model.policyengine.PolicyType
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.model.policyengine.Rule
-import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.model.redshiftdata.StatementCancellationResponse
-import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.model.redshiftdata.StatementExecutionResponse
-import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.model.redshiftdata.StatementExecutionStatus
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.security.DprAuthAwareAuthenticationToken
-import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.service.model.Prompt
-import java.sql.SQLException
 import java.time.LocalDateTime
-import java.util.UUID
 
-class ConfiguredApiServiceTest {
+class SyncDataApiServiceTest {
   private val productDefinitionRepository: ProductDefinitionRepository = JsonFileProductDefinitionRepository(
     listOf("productDefinition.json"),
     DefinitionGsonConfig().definitionGson(IsoLocalDateTimeTypeAdaptor()),
   )
   private val configuredApiRepository: ConfiguredApiRepository = mock<ConfiguredApiRepository>()
-  private val redshiftDataApiRepository: RedshiftDataApiRepository = mock<RedshiftDataApiRepository>()
-  private val athenaApiRepository: AthenaApiRepository = mock<AthenaApiRepository>()
   private val expectedRepositoryResult = listOf(
     mapOf(
       "PRISONNUMBER" to "1",
@@ -103,7 +88,7 @@ class ConfiguredApiServiceTest {
   private val policyEngineResult = "(origin_code='WWI' AND lower(direction)='out') OR (destination_code='WWI' AND lower(direction)='in')"
   private val tableIdGenerator: TableIdGenerator = TableIdGenerator()
   private val datasetHelper: DatasetHelper = DatasetHelper()
-  private val configuredApiService = ConfiguredApiService(productDefinitionRepository, configuredApiRepository, redshiftDataApiRepository, athenaApiRepository, tableIdGenerator, datasetHelper)
+  private val configuredApiService = SyncDataApiService(productDefinitionRepository, configuredApiRepository)
 
   @BeforeEach
   fun setup() {
@@ -544,7 +529,7 @@ class ConfiguredApiServiceTest {
       listOf("productDefinitionPolicyNoAction.json"),
       DefinitionGsonConfig().definitionGson(IsoLocalDateTimeTypeAdaptor()),
     )
-    val configuredApiService = ConfiguredApiService(productDefinitionRepository, configuredApiRepository, redshiftDataApiRepository, athenaApiRepository, tableIdGenerator, datasetHelper)
+    val configuredApiService = SyncDataApiService(productDefinitionRepository, configuredApiRepository)
     whenever(authToken.authorities).thenReturn(listOf(SimpleGrantedAuthority("USER-ROLE-1")))
     val policyEngineResult = "TRUE"
     val reportId = "definition-policy-no-action"
@@ -720,7 +705,7 @@ class ConfiguredApiServiceTest {
     val e = org.junit.jupiter.api.assertThrows<ValidationException> {
       configuredApiService.validateAndFetchData(reportId, reportVariantId, filters, selectedPage, pageSize, sortColumn, sortedAsc, authToken)
     }
-    assertEquals("${ConfiguredApiService.INVALID_REPORT_ID_MESSAGE} $reportId", e.message)
+    assertEquals("${SyncDataApiService.INVALID_REPORT_ID_MESSAGE} $reportId", e.message)
     verify(configuredApiRepository, times(0)).executeQuery(
       any(),
       any(),
@@ -745,7 +730,7 @@ class ConfiguredApiServiceTest {
     val e = org.junit.jupiter.api.assertThrows<ValidationException> {
       configuredApiService.validateAndCount(reportId, reportVariantId, filters, authToken)
     }
-    assertEquals("${ConfiguredApiService.INVALID_REPORT_ID_MESSAGE} $reportId", e.message)
+    assertEquals("${SyncDataApiService.INVALID_REPORT_ID_MESSAGE} $reportId", e.message)
     verify(configuredApiRepository, times(0)).count(any(), any(), any(), any(), any(), any())
   }
 
@@ -761,7 +746,7 @@ class ConfiguredApiServiceTest {
     val e = org.junit.jupiter.api.assertThrows<ValidationException> {
       configuredApiService.validateAndFetchData(reportId, reportVariantId, filters, selectedPage, pageSize, sortColumn, sortedAsc, authToken)
     }
-    assertEquals("${ConfiguredApiService.INVALID_REPORT_VARIANT_ID_MESSAGE} $reportVariantId", e.message)
+    assertEquals("${SyncDataApiService.INVALID_REPORT_VARIANT_ID_MESSAGE} $reportVariantId", e.message)
     verify(configuredApiRepository, times(0)).executeQuery(
       any(),
       any(),
@@ -785,7 +770,7 @@ class ConfiguredApiServiceTest {
     val e = org.junit.jupiter.api.assertThrows<ValidationException> {
       configuredApiService.validateAndCount(reportId, reportVariantId, filters, authToken)
     }
-    assertEquals("${ConfiguredApiService.INVALID_REPORT_VARIANT_ID_MESSAGE} $reportVariantId", e.message)
+    assertEquals("${SyncDataApiService.INVALID_REPORT_VARIANT_ID_MESSAGE} $reportVariantId", e.message)
     verify(configuredApiRepository, times(0)).count(any(), any(), any(), any(), any(), any())
   }
 
@@ -827,7 +812,7 @@ class ConfiguredApiServiceTest {
     val e = org.junit.jupiter.api.assertThrows<ValidationException> {
       configuredApiService.validateAndFetchData(reportId, reportVariantId, filters, selectedPage, pageSize, sortColumn, sortedAsc, authToken)
     }
-    assertEquals(ConfiguredApiService.INVALID_FILTERS_MESSAGE, e.message)
+    assertEquals(SyncDataApiService.INVALID_FILTERS_MESSAGE, e.message)
     verify(configuredApiRepository, times(0)).executeQuery(
       any(),
       any(),
@@ -854,7 +839,7 @@ class ConfiguredApiServiceTest {
     val e = org.junit.jupiter.api.assertThrows<ValidationException> {
       configuredApiService.validateAndFetchData(reportId, reportVariantId, emptyMap(), selectedPage, pageSize, sortColumn, sortedAsc, authToken, setOf(fieldId), "ab")
     }
-    assertEquals(ConfiguredApiService.INVALID_FILTERS_MESSAGE, e.message)
+    assertEquals(SyncDataApiService.INVALID_FILTERS_MESSAGE, e.message)
     verify(configuredApiRepository, times(0)).executeQuery(
       any(),
       any(),
@@ -881,7 +866,7 @@ class ConfiguredApiServiceTest {
     val e = org.junit.jupiter.api.assertThrows<ValidationException> {
       configuredApiService.validateAndFetchData(reportId, reportVariantId, emptyMap(), selectedPage, pageSize, sortColumn, sortedAsc, authToken, setOf(fieldId), "ab")
     }
-    assertEquals(ConfiguredApiService.INVALID_DYNAMIC_FILTER_MESSAGE, e.message)
+    assertEquals(SyncDataApiService.INVALID_DYNAMIC_FILTER_MESSAGE, e.message)
     verify(configuredApiRepository, times(0)).executeQuery(
       any(),
       any(),
@@ -907,7 +892,7 @@ class ConfiguredApiServiceTest {
     val e = org.junit.jupiter.api.assertThrows<ValidationException> {
       configuredApiService.validateAndFetchData(reportId, "last-year", emptyMap(), selectedPage, pageSize, sortColumn, sortedAsc, authToken)
     }
-    assertEquals(ConfiguredApiService.MISSING_MANDATORY_FILTER_MESSAGE + " Date", e.message)
+    assertEquals(SyncDataApiService.MISSING_MANDATORY_FILTER_MESSAGE + " Date", e.message)
   }
 
   @Test
@@ -945,7 +930,7 @@ class ConfiguredApiServiceTest {
     val e = org.junit.jupiter.api.assertThrows<ValidationException> {
       configuredApiService.validateAndFetchData(reportId, "last-year", filters, selectedPage, pageSize, sortColumn, sortedAsc, authToken)
     }
-    assertEquals(ConfiguredApiService.FILTER_VALUE_DOES_NOT_MATCH_PATTERN_MESSAGE + " Invalid [A-Z]{3,3}", e.message)
+    assertEquals(SyncDataApiService.FILTER_VALUE_DOES_NOT_MATCH_PATTERN_MESSAGE + " Invalid [A-Z]{3,3}", e.message)
   }
 
   @Test
@@ -979,7 +964,7 @@ class ConfiguredApiServiceTest {
     val e = org.junit.jupiter.api.assertThrows<ValidationException> {
       configuredApiService.validateAndCount(reportId, "last-year", emptyMap(), authToken)
     }
-    assertEquals(ConfiguredApiService.MISSING_MANDATORY_FILTER_MESSAGE + " Date", e.message)
+    assertEquals(SyncDataApiService.MISSING_MANDATORY_FILTER_MESSAGE + " Date", e.message)
   }
 
   @Test
@@ -991,32 +976,7 @@ class ConfiguredApiServiceTest {
     val e = org.junit.jupiter.api.assertThrows<ValidationException> {
       configuredApiService.validateAndCount(reportId, "last-year", filters, authToken)
     }
-    assertEquals(ConfiguredApiService.FILTER_VALUE_DOES_NOT_MATCH_PATTERN_MESSAGE + " Invalid [A-Z]{3,3}", e.message)
-  }
-
-  @Test
-  fun `validateAndExecuteStatementAsync should throw an exception for a mandatory filter with no value`() {
-    val sortColumn = "date"
-    val sortedAsc = true
-
-    val e = org.junit.jupiter.api.assertThrows<ValidationException> {
-      configuredApiService.validateAndExecuteStatementAsync(reportId, "last-year", emptyMap(), sortColumn, sortedAsc, authToken)
-    }
-    assertEquals(ConfiguredApiService.MISSING_MANDATORY_FILTER_MESSAGE + " Date", e.message)
-  }
-
-  @Test
-  fun `validateAndExecuteStatementAsync should throw an exception for a filter value that does not match the validation pattern`() {
-    val sortColumn = "date"
-    val sortedAsc = true
-    val filters = mapOf(
-      "date.start" to "2000-01-02",
-      "origin" to "Invalid",
-    )
-    val e = org.junit.jupiter.api.assertThrows<ValidationException> {
-      configuredApiService.validateAndExecuteStatementAsync(reportId, "last-year", filters, sortColumn, sortedAsc, authToken)
-    }
-    assertEquals(ConfiguredApiService.FILTER_VALUE_DOES_NOT_MATCH_PATTERN_MESSAGE + " Invalid [A-Z]{3,3}", e.message)
+    assertEquals(SyncDataApiService.FILTER_VALUE_DOES_NOT_MATCH_PATTERN_MESSAGE + " Invalid [A-Z]{3,3}", e.message)
   }
 
   @Test
@@ -1026,7 +986,7 @@ class ConfiguredApiServiceTest {
     val e = org.junit.jupiter.api.assertThrows<ValidationException> {
       configuredApiService.validateAndCount(reportId, reportVariantId, filters, authToken)
     }
-    assertEquals(ConfiguredApiService.INVALID_FILTERS_MESSAGE, e.message)
+    assertEquals(SyncDataApiService.INVALID_FILTERS_MESSAGE, e.message)
     verify(configuredApiRepository, times(0)).count(any(), any(), any(), any(), any(), any())
   }
 
@@ -1041,7 +1001,7 @@ class ConfiguredApiServiceTest {
     val e = org.junit.jupiter.api.assertThrows<ValidationException> {
       configuredApiService.validateAndFetchData(reportId, reportVariantId, filters, selectedPage, pageSize, sortColumn, sortedAsc, authToken)
     }
-    assertEquals(ConfiguredApiService.INVALID_FILTERS_MESSAGE, e.message)
+    assertEquals(SyncDataApiService.INVALID_FILTERS_MESSAGE, e.message)
     verify(configuredApiRepository, times(0)).executeQuery(
       any(),
       any(),
@@ -1064,7 +1024,7 @@ class ConfiguredApiServiceTest {
     val e = org.junit.jupiter.api.assertThrows<ValidationException> {
       configuredApiService.validateAndCount(reportId, reportVariantId, filters, authToken)
     }
-    assertEquals(ConfiguredApiService.INVALID_FILTERS_MESSAGE, e.message)
+    assertEquals(SyncDataApiService.INVALID_FILTERS_MESSAGE, e.message)
     verify(configuredApiRepository, times(0)).count(any(), any(), any(), any(), any(), any())
   }
 
@@ -1079,7 +1039,7 @@ class ConfiguredApiServiceTest {
     val e = org.junit.jupiter.api.assertThrows<ValidationException> {
       configuredApiService.validateAndFetchData(reportId, reportVariantId, filters, selectedPage, pageSize, sortColumn, sortedAsc, authToken)
     }
-    assertEquals(ConfiguredApiService.INVALID_STATIC_OPTIONS_MESSAGE, e.message)
+    assertEquals(SyncDataApiService.INVALID_STATIC_OPTIONS_MESSAGE, e.message)
     verify(configuredApiRepository, times(0)).executeQuery(
       any(),
       any(),
@@ -1102,7 +1062,7 @@ class ConfiguredApiServiceTest {
     val e = org.junit.jupiter.api.assertThrows<ValidationException> {
       configuredApiService.validateAndCount(reportId, reportVariantId, filters, authToken)
     }
-    assertEquals(ConfiguredApiService.INVALID_STATIC_OPTIONS_MESSAGE, e.message)
+    assertEquals(SyncDataApiService.INVALID_STATIC_OPTIONS_MESSAGE, e.message)
     verify(configuredApiRepository, times(0)).count(any(), any(), any(), any(), any(), any())
   }
 
@@ -1117,7 +1077,7 @@ class ConfiguredApiServiceTest {
     val e = org.junit.jupiter.api.assertThrows<ValidationException> {
       configuredApiService.validateAndFetchData(reportId, reportVariantId, filters, selectedPage, pageSize, sortColumn, sortedAsc, authToken)
     }
-    assertEquals(ConfiguredApiService.INVALID_STATIC_OPTIONS_MESSAGE, e.message)
+    assertEquals(SyncDataApiService.INVALID_STATIC_OPTIONS_MESSAGE, e.message)
     verify(configuredApiRepository, times(0)).executeQuery(
       any(),
       any(),
@@ -1154,7 +1114,7 @@ class ConfiguredApiServiceTest {
         "A",
       )
     }
-    assertEquals(ConfiguredApiService.INVALID_DYNAMIC_OPTIONS_MESSAGE, e.message)
+    assertEquals(SyncDataApiService.INVALID_DYNAMIC_OPTIONS_MESSAGE, e.message)
     verify(configuredApiRepository, times(0)).executeQuery(
       any(),
       any(),
@@ -1177,7 +1137,7 @@ class ConfiguredApiServiceTest {
     val e = org.junit.jupiter.api.assertThrows<ValidationException> {
       configuredApiService.validateAndCount(reportId, reportVariantId, filters, authToken)
     }
-    assertEquals(ConfiguredApiService.INVALID_STATIC_OPTIONS_MESSAGE, e.message)
+    assertEquals(SyncDataApiService.INVALID_STATIC_OPTIONS_MESSAGE, e.message)
     verify(configuredApiRepository, times(0)).count(any(), any(), any(), any(), any(), any())
   }
 
@@ -1318,7 +1278,7 @@ class ConfiguredApiServiceTest {
       mapOf("9" to "1"),
     )
     val productDefRepo = mock<ProductDefinitionRepository>()
-    val configuredApiService = ConfiguredApiService(productDefRepo, configuredApiRepository, redshiftDataApiRepository, athenaApiRepository, tableIdGenerator, datasetHelper)
+    val configuredApiService = SyncDataApiService(productDefRepo, configuredApiRepository)
     val dataSourceName = "name"
 
     whenever(productDefRepo.getProductDefinitions())
@@ -1440,332 +1400,5 @@ class ConfiguredApiServiceTest {
       productDefinition = singleReportProductDefinition,
     )
     assertEquals(expectedServiceResult, actual)
-  }
-
-  @Test
-  fun `should make the async call to the RedshiftDataApiRepository for datamart with all provided arguments when validateAndExecuteStatementAsync is called`() {
-    val configuredApiService = ConfiguredApiService(productDefinitionRepository, configuredApiRepository, redshiftDataApiRepository, athenaApiRepository, tableIdGenerator, datasetHelper)
-    val filters = mapOf("is_closed" to "true", "date$RANGE_FILTER_START_SUFFIX" to "2023-04-25", "date$RANGE_FILTER_END_SUFFIX" to "2023-09-10")
-    val repositoryFilters = listOf(Filter("is_closed", "true", BOOLEAN), Filter("date", "2023-04-25", DATE_RANGE_START), Filter("date", "2023-09-10", DATE_RANGE_END))
-    val sortColumn = "date"
-    val sortedAsc = true
-    val productDefinition = productDefinitionRepository.getProductDefinitions().first()
-    val singleReportProductDefinition = productDefinitionRepository.getSingleReportProductDefinition(productDefinition.id, productDefinition.report.first().id)
-    val executionId = UUID.randomUUID().toString()
-    val tableId = executionId.replace("-", "_")
-    val statementExecutionResponse = StatementExecutionResponse(tableId, executionId)
-    whenever(
-      redshiftDataApiRepository.executeQueryAsync(
-        productDefinition = singleReportProductDefinition,
-        filters = repositoryFilters,
-        sortColumn = sortColumn,
-        sortedAsc = sortedAsc,
-        policyEngineResult = policyEngineResult,
-        prompts = emptyList(),
-        userToken = authToken,
-      ),
-    ).thenReturn(statementExecutionResponse)
-
-    val actual = configuredApiService.validateAndExecuteStatementAsync(reportId, reportVariantId, filters, sortColumn, sortedAsc, authToken)
-
-    verify(redshiftDataApiRepository, times(1)).executeQueryAsync(
-      productDefinition = singleReportProductDefinition,
-      filters = repositoryFilters,
-      sortColumn = sortColumn,
-      sortedAsc = sortedAsc,
-      policyEngineResult = policyEngineResult,
-      prompts = emptyList(),
-      userToken = authToken,
-    )
-    assertEquals(statementExecutionResponse, actual)
-  }
-
-  @ParameterizedTest
-  @CsvSource(
-    "nomis_db, nomis, productDefinitionNomis.json",
-    "bodmis_db, bodmis, productDefinitionBodmis.json",
-  )
-  fun `should make the async call to the AthenaApiRepository for nomis and bodmis with all provided arguments when validateAndExecuteStatementAsync is called`(database: String, catalog: String, definitionFile: String) {
-    val productDefinitionRepository: ProductDefinitionRepository = JsonFileProductDefinitionRepository(
-      listOf(definitionFile),
-      DefinitionGsonConfig().definitionGson(IsoLocalDateTimeTypeAdaptor()),
-    )
-    val configuredApiService = ConfiguredApiService(productDefinitionRepository, configuredApiRepository, redshiftDataApiRepository, athenaApiRepository, tableIdGenerator, datasetHelper)
-    val filters = mapOf("is_closed" to "true", "date$RANGE_FILTER_START_SUFFIX" to "2023-04-25", "date$RANGE_FILTER_END_SUFFIX" to "2023-09-10")
-    val repositoryFilters = listOf(Filter("is_closed", "true", BOOLEAN), Filter("date", "2023-04-25", DATE_RANGE_START), Filter("date", "2023-09-10", DATE_RANGE_END))
-    val sortColumn = "date"
-    val sortedAsc = true
-    val productDefinition = productDefinitionRepository.getProductDefinitions().first()
-    val singleReportProductDefinition = productDefinitionRepository.getSingleReportProductDefinition(productDefinition.id, productDefinition.report.first().id)
-    val executionId = UUID.randomUUID().toString()
-    val tableId = executionId.replace("-", "_")
-    val statementExecutionResponse = StatementExecutionResponse(tableId, executionId)
-    whenever(
-      athenaApiRepository.executeQueryAsync(
-        productDefinition = singleReportProductDefinition,
-        filters = repositoryFilters,
-        sortColumn = sortColumn,
-        sortedAsc = sortedAsc,
-        policyEngineResult = policyEngineResult,
-        prompts = emptyList(),
-        userToken = authToken,
-      ),
-    ).thenReturn(statementExecutionResponse)
-
-    val actual = configuredApiService.validateAndExecuteStatementAsync(reportId, reportVariantId, filters, sortColumn, sortedAsc, authToken)
-
-    verify(athenaApiRepository, times(1)).executeQueryAsync(
-      productDefinition = singleReportProductDefinition,
-      filters = repositoryFilters,
-      sortColumn = sortColumn,
-      sortedAsc = sortedAsc,
-      policyEngineResult = policyEngineResult,
-      prompts = emptyList(),
-      userToken = authToken,
-    )
-    verifyNoInteractions(redshiftDataApiRepository)
-    assertEquals(statementExecutionResponse, actual)
-  }
-
-  @Test
-  fun `should call the RedshiftDataApiRepository for datamart with the statement execution ID when getStatementStatus is called`() {
-    val configuredApiService = ConfiguredApiService(productDefinitionRepository, configuredApiRepository, redshiftDataApiRepository, athenaApiRepository, tableIdGenerator, datasetHelper)
-    val statementId = "statementId"
-    val status = "FINISHED"
-    val duration = 278109264L
-    val resultRows = 10L
-    val resultSize = 100L
-    val statementExecutionStatus = StatementExecutionStatus(
-      status,
-      duration,
-      resultRows,
-      resultSize,
-    )
-    whenever(
-      redshiftDataApiRepository.getStatementStatus(statementId),
-    ).thenReturn(statementExecutionStatus)
-
-    val actual = configuredApiService.getStatementStatus(statementId, "external-movements", "last-month")
-    verify(redshiftDataApiRepository, times(1)).getStatementStatus(statementId)
-    verifyNoInteractions(athenaApiRepository)
-    assertEquals(statementExecutionStatus, actual)
-  }
-
-  @ParameterizedTest
-  @ValueSource(strings = ["productDefinitionNomis.json", "productDefinitionBodmis.json"])
-  fun `should call the AthenaApiRepository for nomis and bodmis with the statement execution ID when getStatementStatus is called`(definitionFile: String) {
-    val productDefinitionRepository: ProductDefinitionRepository = JsonFileProductDefinitionRepository(
-      listOf(definitionFile),
-      DefinitionGsonConfig().definitionGson(IsoLocalDateTimeTypeAdaptor()),
-    )
-    val configuredApiService = ConfiguredApiService(productDefinitionRepository, configuredApiRepository, redshiftDataApiRepository, athenaApiRepository, tableIdGenerator, datasetHelper)
-    val statementId = "statementId"
-    val status = "FINISHED"
-    val duration = 278109264L
-    val resultRows = 10L
-    val resultSize = 100L
-    val statementExecutionStatus = StatementExecutionStatus(
-      status,
-      duration,
-      resultRows,
-      resultSize,
-    )
-    whenever(
-      athenaApiRepository.getStatementStatus(statementId),
-    ).thenReturn(statementExecutionStatus)
-
-    val actual = configuredApiService.getStatementStatus(statementId, "external-movements", "last-month")
-    verify(athenaApiRepository, times(1)).getStatementStatus(statementId)
-    verifyNoInteractions(redshiftDataApiRepository)
-    assertEquals(statementExecutionStatus, actual)
-  }
-
-  @Test
-  fun `should call the RedshiftDataApiRepository for datamart with the statement execution ID when cancelStatementExecution is called`() {
-    val configuredApiService = ConfiguredApiService(productDefinitionRepository, configuredApiRepository, redshiftDataApiRepository, athenaApiRepository, tableIdGenerator, datasetHelper)
-    val statementId = "statementId"
-    val statementCancellationResponse = StatementCancellationResponse(
-      true,
-    )
-    whenever(
-      redshiftDataApiRepository.cancelStatementExecution(statementId),
-    ).thenReturn(statementCancellationResponse)
-
-    val actual = configuredApiService.cancelStatementExecution(statementId, "external-movements", "last-month")
-    verify(redshiftDataApiRepository, times(1)).cancelStatementExecution(statementId)
-    verifyNoInteractions(athenaApiRepository)
-    assertEquals(statementCancellationResponse, actual)
-  }
-
-  @ParameterizedTest
-  @ValueSource(strings = ["productDefinitionNomis.json", "productDefinitionBodmis.json"])
-  fun `should call the AthenaApiRepository for nomis and bodmis with the statement execution ID when cancelStatementExecution is called`(definitionFile: String) {
-    val productDefinitionRepository: ProductDefinitionRepository = JsonFileProductDefinitionRepository(
-      listOf(definitionFile),
-      DefinitionGsonConfig().definitionGson(IsoLocalDateTimeTypeAdaptor()),
-    )
-    val configuredApiService = ConfiguredApiService(productDefinitionRepository, configuredApiRepository, redshiftDataApiRepository, athenaApiRepository, tableIdGenerator, datasetHelper)
-    val statementId = "statementId"
-    val statementCancellationResponse = StatementCancellationResponse(true)
-    whenever(
-      athenaApiRepository.cancelStatementExecution(statementId),
-    ).thenReturn(statementCancellationResponse)
-
-    val actual = configuredApiService.cancelStatementExecution(statementId, "external-movements", "last-month")
-    verify(athenaApiRepository, times(1)).cancelStatementExecution(statementId)
-    verifyNoInteractions(redshiftDataApiRepository)
-    assertEquals(statementCancellationResponse, actual)
-  }
-
-  @Test
-  fun `validateAndExecuteStatementAsync should not fail validation for filters which were converted from DPD parameters and convert these filters to prompts`() {
-    val productDefinitionRepository: ProductDefinitionRepository = JsonFileProductDefinitionRepository(
-      listOf("productDefinitionWithParameters.json"),
-      DefinitionGsonConfig().definitionGson(IsoLocalDateTimeTypeAdaptor()),
-    )
-    val productDefinition = productDefinitionRepository.getProductDefinitions().first()
-    val singleReportProductDefinition = productDefinitionRepository.getSingleReportProductDefinition(productDefinition.id, productDefinition.report.first().id)
-    val configuredApiService = ConfiguredApiService(productDefinitionRepository, configuredApiRepository, redshiftDataApiRepository, athenaApiRepository, tableIdGenerator, datasetHelper)
-    val parameterName = "prisoner_number"
-    val parameterValue = "somePrisonerNumber"
-    val filters = mapOf(
-      "origin" to "someOrigin",
-      parameterName to parameterValue,
-    )
-    val repositoryFilters = listOf(Filter("origin", "someOrigin", STANDARD))
-    val prompts = listOf(Prompt(parameterName, parameterValue, FilterType.Text))
-    val sortColumn = "date"
-    val sortedAsc = true
-    val executionId = UUID.randomUUID().toString()
-    val tableId = executionId.replace("-", "_")
-    val statementExecutionResponse = StatementExecutionResponse(tableId, executionId)
-    whenever(
-      redshiftDataApiRepository.executeQueryAsync(
-        productDefinition = singleReportProductDefinition,
-        filters = repositoryFilters,
-        sortColumn = sortColumn,
-        sortedAsc = sortedAsc,
-        policyEngineResult = policyEngineResult,
-        prompts = prompts,
-        userToken = authToken,
-      ),
-    ).thenReturn(statementExecutionResponse)
-
-    val actual = configuredApiService.validateAndExecuteStatementAsync("external-movements-with-parameters", reportVariantId, filters, sortColumn, sortedAsc, authToken)
-
-    verify(redshiftDataApiRepository, times(1)).executeQueryAsync(
-      productDefinition = singleReportProductDefinition,
-      filters = repositoryFilters,
-      sortColumn = sortColumn,
-      sortedAsc = sortedAsc,
-      policyEngineResult = policyEngineResult,
-      prompts = prompts,
-      userToken = authToken,
-    )
-    assertEquals(statementExecutionResponse, actual)
-  }
-
-  @Test
-  fun `should call the repository with all provided arguments when getStatementResult is called`() {
-    val tableId = TableIdGenerator().generateNewExternalTableId()
-    val selectedPage = 1L
-    val pageSize = 20L
-    whenever(
-      redshiftDataApiRepository.getPaginatedExternalTableResult(tableId, selectedPage, pageSize),
-    ).thenReturn(expectedRepositoryResult)
-
-    val actual = configuredApiService.getStatementResult(
-      tableId,
-      reportId,
-      reportVariantId,
-      selectedPage = selectedPage,
-      pageSize = pageSize,
-    )
-
-    assertEquals(expectedServiceResult, actual)
-  }
-
-  @Test
-  fun `getStatementResult should apply formulas to the rows returned by the repository`() {
-    val selectedPage = 1L
-    val pageSize = 20L
-    val expectedRepositoryResult = listOf(
-      mapOf("PRISONNUMBER" to "1", "NAME" to "FirstName", "ORIGIN" to "OriginLocation", "ORIGIN_CODE" to "abc"),
-    )
-    val expectedServiceResult = listOf(
-      mapOf("prisonNumber" to "1", "name" to "FirstName", "origin" to "OriginLocation", "origin_code" to "OriginLocation"),
-    )
-    val productDefinitionRepository: ProductDefinitionRepository = JsonFileProductDefinitionRepository(
-      listOf("productDefinitionWithFormula.json"),
-      DefinitionGsonConfig().definitionGson(IsoLocalDateTimeTypeAdaptor()),
-    )
-    val configuredApiService = ConfiguredApiService(productDefinitionRepository, configuredApiRepository, redshiftDataApiRepository, athenaApiRepository, tableIdGenerator, datasetHelper)
-    val executionID = UUID.randomUUID().toString()
-    whenever(
-      redshiftDataApiRepository.getPaginatedExternalTableResult(executionID, selectedPage, pageSize),
-    ).thenReturn(expectedRepositoryResult)
-
-    val actual = configuredApiService.getStatementResult(
-      tableId = executionID,
-      reportId = reportId,
-      reportVariantId = reportVariantId,
-      selectedPage = selectedPage,
-      pageSize = pageSize,
-    )
-
-    assertEquals(expectedServiceResult, actual)
-  }
-
-  @Test
-  fun `should call the repository with all provided arguments when getSummaryResult is called`() {
-    val tableId = TableIdGenerator().generateNewExternalTableId()
-    val summaryId = "summaryId"
-    whenever(
-      redshiftDataApiRepository.getFullExternalTableResult(tableIdGenerator.getTableSummaryId(tableId, summaryId)),
-    ).thenReturn(listOf(mapOf("TOTAL" to 1)))
-
-    val actual = configuredApiService.getSummaryResult(
-      tableId,
-      summaryId,
-      reportId,
-      reportVariantId,
-    )
-
-    assertEquals(listOf(mapOf("total" to 1)), actual)
-  }
-
-  @Test
-  fun `should create and query summary table when it doesn't exist`() {
-    val tableId = TableIdGenerator().generateNewExternalTableId()
-    val summaryId = "summaryId"
-    whenever(
-      redshiftDataApiRepository.getFullExternalTableResult(tableIdGenerator.getTableSummaryId(tableId, summaryId)),
-    )
-      .thenThrow(UncategorizedSQLException("EntityNotFoundException from glue - Entity Not Found", "", SQLException()))
-      .thenReturn(listOf(mapOf("TOTAL" to 1)))
-
-    val actual = configuredApiService.getSummaryResult(
-      tableId,
-      summaryId,
-      reportId,
-      reportVariantId,
-    )
-
-    assertEquals(listOf(mapOf("total" to 1)), actual)
-    verify(redshiftDataApiRepository, times(2)).getFullExternalTableResult(any(), anyOrNull())
-    verify(configuredApiRepository).createSummaryTable(any(), any(), any(), any())
-  }
-
-  @Test
-  fun `should call the repository with all provided arguments when count is called`() {
-    val tableId = "123"
-    val expectedRepositoryResult = 5L
-    whenever(
-      redshiftDataApiRepository.count(tableId),
-    ).thenReturn(expectedRepositoryResult)
-
-    val actual = configuredApiService.count(tableId)
-
-    assertEquals(Count(expectedRepositoryResult), actual)
   }
 }
